+++ spec/Pim/Bundle/CatalogBundle/Manager/FamilyManagerSpec.php
@@ -32,12 +32,12 @@ class FamilyManagerSpec extends ObjectBehavior
 
     function it_is_a_saver()
     {
-        $this->shouldHaveType('Pim\Component\Resource\Model\SaverInterface');
+        $this->shouldHaveType('Akeneo\Component\Persistence\SaverInterface');
     }
 
     function it_is_a_remover()
     {
-        $this->shouldHaveType('Pim\Component\Resource\Model\RemoverInterface');
+        $this->shouldHaveType('Akeneo\Component\Persistence\RemoverInterface');
     }
 
     function it_provides_a_choice_list($userContext, $repository)