+++ src/Pim/Bundle/EnrichBundle/Controller/CurrencyController.php
@@ -4,7 +4,7 @@ namespace Pim\Bundle\EnrichBundle\Controller;
 
 use Doctrine\Common\Persistence\ManagerRegistry;
 use Oro\Bundle\SecurityBundle\Annotation\AclAncestor;
-use Pim\Component\Resource\Model\SaverInterface;
+use Akeneo\Component\Persistence\SaverInterface;
 use Pim\Bundle\CatalogBundle\Entity\Currency;
 use Pim\Bundle\EnrichBundle\AbstractController\AbstractDoctrineController;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;